Add clong::run helper
